moved sorter dir from config to helper
