+++ src/civitdl/batch_download.py
@@ -6,7 +6,7 @@ from typing import Dict
 from .get_model import download_model
 from .helper.utils import run_in_dev
 
-from civitdl.config.sorter import basic, tags
+from civitdl.helper.sorter import basic, tags
 
 
 def choose_sorter(sorter: Dict):
